# user-manager/user-manager-rest-api/openapi/user-manager-rest-api.yml
...
openapi: 3.0.0
info:

  # arbitrary string that specifies the version of the API
  version: 0.0.1

  # API name
  title: Butterfly User Manager API

  # Extended information about the API
  description: >-
    Butterfly User Manager REST API

  contact:
    name: DStack Group
    email: dstackgroup@gmail.com
    # url: https://example.com/

# API server and base URL
servers:
  - url: http://localhost:5000
    description: Development Server

# tags are used for organizing operations
...
paths:
...
  /users/{email}/contacts/{contactService}:
    post:
      operationId: createUserContact
      summary: >-
        Creates a new association between an existing user and a new contact platform where the user wishes
        to receive event notifications
      tags: [ 'USER-CONTACTS' ]
      parameters:
        - in: path
          name: email
          schema:
            type: string
            format: email
          required: true
        - in: path
          name: contactService
          schema:
            $ref: '#/components/schemas/ThirdPartyContactService'
          required: true
      requestBody:
        description: Contact account identifier for the given contact service
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contactRef:
                  type: string
                  description: >-
                    the contact account id, such as the Telegram's bot chat id, the Slack's user id,
                    or the Email address (possibly different from the currently registered email)
      responses:
        '201':
          description: OK, new contact info added to the user identified by the given email
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserContact'

        '404':
          description: Error, the user does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message

        '409':
          description: Conflict error, there is already a contact defined for the given user and service
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message

        '422':
          description: Error, the chosen contact service isn't supported
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message

    put:
      operationId: editUserContact
      summary: >-
        Edits an existing association between a user and a contact platform where the user wishes
        to receive event notifications
      tags: [ 'USER-CONTACTS' ]
      parameters:
        - in: path
          name: email
          schema:
            type: string
            format: email
          required: true
        - in: path
          name: contactService
          schema:
            $ref: '#/components/schemas/ThirdPartyContactService'
          required: true
      requestBody:
        description: Updated contact account identifier for the given contact service
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contactRef:
                  type: string
                  description: >-
                    the new contact account id, such as the Telegram's bot chat id, the Slack's user id,
                    or the Email address (possibly different from the currently registered email)
      responses:
        '200':
          description: OK, new contact info updated for the user identified by the given email
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserContact'

        '404':
          description: Error, the user or the given contact association does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message

        '422':
          description: Error, the chosen contact service isn't supported
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message

    delete:
      operationId: deleteUserContact
      summary: Removes the user contact specified by contact service. The user is identified by its email
      tags: [ 'USER-CONTACTS' ]
      parameters:
        - in: path
          name: email
          schema:
            type: string
            format: email
          required: true
        - in: path
          name: contactService
          schema:
            $ref: '#/components/schemas/ThirdPartyContactService'
          required: true
      responses:
        '200':
          description: OK, contact info deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    nullable: true
        '404':
          description: Error, the contact info does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message
  /projects:
    get:
      operationId: getProjects
      summary: Retrieves the list of registered projects
      tags: [ 'PROJECTS' ]
      responses:
        '200':
          description: OK, returns a list of projects records.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'

    post:
      operationId: createProject
      summary: Creates new project and returns the project record
      tags: [ 'PROJECTS' ]
      requestBody:
        description: Project record
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProject'
      responses:
        '201':
          description: OK, project inserted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '409':
          description: Unique constraint violation, another project exists with the same name
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message
  /projects/{projectName}:
    get:
      operationId: getProjectByName
      summary: Searches a project by its name and returns its record data
      tags: [ 'PROJECTS' ]
      parameters:
        - in: path
          name: projectName
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK, project found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '404':
          description: Error, the project does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message

    put:
      operationId: updateProjectByName
      summary: Updates the data of the project identified by the given project name and returns the updated project record
      tags: [ 'PROJECTS' ]
      parameters:
      - in: path
        name: projectName
        schema:
          type: string
        required: true
      requestBody:
        description: Project record
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProject'
      responses:
        '200':
          description: OK, project updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '404':
          description: Error, the project does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message

    delete:
      operationId: deleteProjectByName
      summary: Removes the project that is registered with the given project name
      tags: [ 'PROJECTS' ]
      parameters:
        - in: path
          name: projectName
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK, project deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    nullable: true
        '404':
          description: Error, the project does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message
...
components:
  schemas:
...
    ThirdPartyContactService:
      type: string
      enum:
        - TELEGRAM
        - SLACK
        - EMAIL
...
    UserContact:
      type: object
      required:
        - userContactid
        - userId
        - contactType
        - contactRef
      properties:
        userContactId:
          type: string
        userId:
          type: string
        contactType:
          $ref: '#/components/schemas/ThirdPartyContactService'
        contactRef:
          type: string
      example:
        userContactId: "23"
        userId: "1"
        contactType: "TELEGRAM"
        contactRef: "12739814"
...
    ProjectURL:
      type: object
      properties:
        REDMINE:
          type: string
          format: url
          nullable: true
        GITLAB:
          type: string
          format: url
          nullable: true
        SONARQUBE:
          type: string
          format: url
          nullable: true
      example:
        REDMINE: "http://redmine.dstack.unipd.it"
        GITLAB: "http://gitlab.dstack.unipd.it"
    CreateProject:
      type: object
      required:
        - projectName
        - projectURL
      properties:
        projectName:
          type: string
        projectURL:
          $ref: '#/components/schemas/ProjectURL'
      example:
        projectName: "Butterfly"
        projectURL:
          REDMINE: "http://redmine.dstack.unipd.it"
          GITLAB: "http://gitlab.dstack.unipd.it"
    UpdateProject:
      type: object
      required:
        - projectName
        - projectURL
      properties:
        projectName:
          type: string
        projectURL:
          $ref: '#/components/schemas/ProjectURL'
      example:
        projectName: "Butterfly"
        projectURL:
          SONARQUBE: "http://sonarqube.dstack.unipd.it"
    Project:
      type: object
      required:
        - projectId
        - projectName
        - projectURL
        - created
      properties:
        projectId:
          type: string
        projectName:
          type: string
        projectURL:
          $ref: '#/components/schemas/ProjectURL'
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
          nullable: true
      example:
        projectId: "2"
        projectName: "Butterfly"
        projectURL:
          REDMINE: "http://redmine.dstack.unipd.it"
          GITLAB: "http://gitlab.dstack.unipd.it"
          SONARQUBE: "http://sonarqube.dstack.unipd.it"
...